requirements.txt
# SmallTree Academy Requirements
# Core Flask Framework
Flask==2.0.3
Flask-WTF==0.15.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==3.1.0
SQLAlchemy==2.0.23
Jinja2==3.0.3
Werkzeug==2.0.3

# Production Server (removed invalid package)

# Forms & Validation
WTForms==3.0.1
email_validator==1.3.1

# Environment & Utils
python-dotenv==0.19.2

# Document Processing (Optional)
python-docx==1.0.1
openpyxl==3.1.3
htmldocx>=0.0.6  # For HTML to Word conversion

# Cloudflare R2 Storage
boto3==1.34.19  # AWS SDK (compatible with R2)
Pillow==10.0.0  # Image processing

# AI Services (Optional - install separately if needed)
google-generativeai>=0.3.0
openai>=0.28.0
requests>=2.28.0
cohere>=5.0.0
groq>=0.4.0
anthropic>=0.20.0